Seperated chart logic from UI
