+++ src/logic/Chart-Data/ChartData.js
@@ -1,0 +1,68 @@
+import { getStyle, hexToRgba } from '@coreui/coreui/dist/js/coreui-utilities';
+
+const brandSuccess = getStyle('--success')
+const brandInfo = getStyle('--info')
+const brandDanger = getStyle('--danger')
+
+//Random Numbers
+function random(min, max) {
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
+
+function getData(set) {
+    var elements = 27;
+    var data1 = [];
+    var data2 = [];
+    var data3 = [];
+
+    for (var i = 0; i <= elements; i++) {
+      data1.push(random(50, 200));
+      data2.push(random(80, 100));
+      data3.push(65);
+    }
+
+    switch (set) {
+      case 1: return data1;
+      case 2: return data2;
+      default: return data3;
+    }
+}
+
+function chartData() {
+    let data1 = getData(1);
+    let data2 = getData(2);
+    let data3 = getData(3);
+    const mainChart = {
+      labels: ['Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa', 'Su'],
+      datasets: [
+        {
+          label: 'My First dataset',
+          backgroundColor: hexToRgba(brandInfo, 10),
+          borderColor: brandInfo,
+          pointHoverBackgroundColor: '#fff',
+          borderWidth: 2,
+          data: data1,
+        },
+        {
+          label: 'My Second dataset',
+          backgroundColor: 'transparent',
+          borderColor: brandSuccess,
+          pointHoverBackgroundColor: '#fff',
+          borderWidth: 2,
+          data: data2,
+        },
+        {
+          label: 'My Third dataset',
+          backgroundColor: 'transparent',
+          borderColor: brandDanger,
+          pointHoverBackgroundColor: '#fff',
+          borderWidth: 1,
+          borderDash: [8, 5],
+          data: data3,
+        },
+      ],
+    };
+    return mainChart;
+}
+
+export default chartData;
